+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 7.4
 - Add Course Widget configuration option: Popup window or Pull-Down in Configuration.php & Help_en.php
 - Add Course Widget: select / Pull-Down in Widgets.fnc.php
 - Update French & Spanish translations in rosariosis.po, help.po
+- Add Total Credits in ReportCards.fnc.php
 
 Changes in 7.3.1
 ----------------